Upload Skill/certificate csv done

+++ src/main/java/com/teksystems/skillportal/controller/RoleController.java
@@ -8,10 +8,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.util.List;
 
 @RestController
-@RequestMapping(value="/role",method= {RequestMethod.GET,RequestMethod.POST,RequestMethod.PUT,RequestMethod.DELETE})
+@RequestMapping(value = "/role", method = {RequestMethod.GET, RequestMethod.POST, RequestMethod.PUT, RequestMethod.DELETE})
 @CrossOrigin("*")
 public class RoleController {
     private static Logger logger = Logger.getLogger(RoleController.class);
@@ -19,24 +20,25 @@ public class RoleController {
     private TokenValidationService tokenValidator;
     @Autowired
     RoleService roleService;
+
     /*
-    * Controller to fetch all the admin Roles of the app to provide the auth in the frontend
-    */
+     * Controller to fetch all the admin Roles of the app to provide the auth in the frontend
+     */
     @GetMapping("/adminRoles")
-    public List<AdminRoles> getAllAdminRoles(HttpServletRequest request){
+    public List<AdminRoles> getAllAdminRoles(HttpServletRequest request) {
         logger.info("adminRoles API Called");
         String employeeId = null;
         List<AdminRoles> toReturn = null;
         try {
             logger.info("Trying to Fetch the Employee Id from the HTTP HEADERS");
             if (!(((HttpServletRequest) request).getHeader("Authorization").toString().equals(null))) {
-                    employeeId = tokenValidator.ExtractEmployeeId(request);
-                    if(employeeId !=null) {
-                        logger.debug("Paramater received : employeeId " + employeeId);
-                        toReturn = roleService.getAdminRoles();
-                    }else{
-                        logger.info("Employee Email id not Present");
-                    }
+                employeeId = tokenValidator.ExtractEmployeeId(request);
+                if (employeeId != null) {
+                    logger.debug("Paramater received : employeeId " + employeeId);
+                    toReturn = roleService.getAdminRoles();
+                } else {
+                    logger.info("Employee Email id not Present");
+                }
 
             } else {
                 logger.info("Authorization Header not found");
@@ -48,23 +50,27 @@ public class RoleController {
     }
 
     /*
-    * Controller for adding new Role.
+     * Controller for adding new Role.
      */
     @PostMapping("/addAdminRole")
-    private void addAdminRole(HttpServletRequest request,@RequestBody AdminRoles role){
+    private void addAdminRole(HttpServletRequest request, @RequestBody AdminRoles role, HttpServletResponse response) {
         logger.info("adminRoles API Called");
         String employeeId = null;
         try {
             logger.info("Trying to Fetch the Employee Id from the HTTP HEADERS");
             if (!(((HttpServletRequest) request).getHeader("Authorization").toString().equals(null))) {
-                employeeId = tokenValidator.ExtractEmployeeId(request);
-                if(employeeId !=null) {
-                    logger.debug("Paramater received : employeeId " + employeeId);
-                    roleService.addRoles(role);
-                }else{
-                    logger.info("Employee Email id not Present");
+                if (tokenValidator.validateAdminRole(request, response)) {
+                    employeeId = tokenValidator.ExtractEmployeeId(request);
+                    if (employeeId != null) {
+                        logger.debug("Paramater received : employeeId " + employeeId);
+                        roleService.addRoles(role);
+                    } else {
+                        logger.info("Employee Email id not Present");
+                    }
+                } else {
+                    logger.debug("Employee doesn't have Admin Role");
+                    ((HttpServletResponse) response).sendError(HttpServletResponse.SC_UNAUTHORIZED, "Invalid Token");
                 }
-
             } else {
                 logger.info("Authorization Header not found");
             }
@@ -74,20 +80,24 @@ public class RoleController {
     }
 
     @DeleteMapping("/deleteRole")
-    public void deleteRole(HttpServletRequest request,@RequestParam String id){
+    public void deleteRole(HttpServletRequest request, @RequestParam String id, HttpServletResponse response) {
         logger.info("adminRoles API Called");
         String employeeId = null;
         try {
             logger.info("Trying to Fetch the Employee Id from the HTTP HEADERS");
             if (!(((HttpServletRequest) request).getHeader("Authorization").toString().equals(null))) {
-                employeeId = tokenValidator.ExtractEmployeeId(request);
-                if(employeeId !=null) {
-                    logger.debug("Paramater received : employeeId " + employeeId);
-                    roleService.deleteRole(id);
-                }else{
-                    logger.info("Employee Email id not Present");
+                if (tokenValidator.validateAdminRole(request, response)) {
+                    employeeId = tokenValidator.ExtractEmployeeId(request);
+                    if (employeeId != null) {
+                        logger.debug("Paramater received : employeeId " + employeeId);
+                        roleService.deleteRole(id);
+                    } else {
+                        logger.info("Employee Email id not Present");
+                    }
+                } else {
+                    logger.debug("Employee doesn't have Admin Role");
+                    ((HttpServletResponse) response).sendError(HttpServletResponse.SC_UNAUTHORIZED, "Invalid Token");
                 }
-
             } else {
                 logger.info("Authorization Header not found");
             }
